utils/import_mails/requirements.txt
====
    Copyright (C) 2007-2023 - LINAGORA
    This program is free software: you can redistribute it and/or modify
    it under the terms of the GNU Affero General Public License as published by
    the Free Software Foundation, either version 3 of the License, or
    (at your option) any later version.

    This program is distributed in the hope that it will be useful,
    but WITHOUT ANY WARRANTY; without even the implied warranty of
    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
    GNU Affero General Public License for more details.

    You should have received a copy of the GNU Affero General Public License
    along with this program.  If not, see <http://www.gnu.org/licenses/>.
====

aenum==2.0.8
certifi==2018.11.29
chardet==3.0.4
idna==2.8
psycopg2-binary==2.7.6.1
PyPika==0.18.4
requests==2.21.0
urllib3==1.25.2
